Extract API stubs into separate file.
This is to keep all API related stubs in one file. Ideally it should be
wrapped into a module but I couldn't make it work.
